chore: Migration AGP to 9.0.0
- AGP(Android Gradle Plugin): 8.13.0 -> 9.0.0
- Compose: 1.10.0 -> 1.10.1
